some really cool (I mean it!) shell logic which fixes bsiegert@s problem - he didn't have an existing /bin/mksh to begin with, so the first mv failed due to "set -e", and thus the && mv mksh.tmp mksh was never executed - there are other ways to fix, but this, devised almost after I saw the problem (but was too ill to check for the problem), is a cool one-liner which fits into the system and the way "we" write Makefiles, too.

+++ ports/Setup.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# $MirOS: ports/Setup.sh,v 1.7 2005/09/12 23:25:55 tg Exp $
+# $MirOS: ports/Setup.sh,v 1.8 2005/09/13 10:52:35 tg Exp $
 #-
 # Copyright (c) 2005
 #	Thorsten "mirabile" Glaser <tg@66h.42h.de>
@@ -338,7 +338,7 @@ fi
 # Install mksh
 set -e
 install -c -s -m 555 mksh $MKSH.$$.1
-mv $MKSH $MKSH.$$.2 && mv $MKSH.$$.1 $MKSH
+test ! -s $MKSH || mv $MKSH $MKSH.$$.2 && mv $MKSH.$$.1 $MKSH
 set +e
 rm -f $MKSH.$$.2
 test -f $MKSH.$$.2 && if mv $MKSH.$$.2 /tmp/deleteme.$$.$RANDOM; then
